--- karma.conf.ci.js
@@ -1,25 +1,0 @@
-module.exports = function(config) {
-	var configuration = {
-		browsers: ['Firefox'],
-		customLaunchers: {
-			Chrome_travis_ci: {
-				base: 'Chrome',
-				flags: ['--no-sandbox']
-			}
-		},
-		frameworks: ['browserify', 'jasmine'],
-		reporters: ['progress', 'html'],
-		preprocessors: {
-			'src/**/*.js': ['browserify']
-		},
-		browserify: {
-			debug: true
-		}
-	};
-
-	if (process.env.TRAVIS) {
-		configuration.browsers.push('Chrome_travis_ci');
-	}
-
-	config.set(configuration);
-};
