+++ test/specs/controller.bar.tests.js
