+++ test/specs/controller.bubble.tests.js
