+++ test/specs/core.controller.tests.js
