+++ test/specs/core.helpers.tests.js
