+++ test/specs/core.interaction.tests.js
