+++ test/specs/core.layoutService.tests.js
