+++ test/specs/core.legend.tests.js
