+++ test/specs/core.plugin.tests.js
