+++ test/specs/core.scaleService.tests.js
