+++ test/specs/core.title.tests.js
