+++ test/specs/core.tooltip.tests.js
