+++ test/specs/element.point.tests.js
