+++ test/specs/element.rectangle.tests.js
