+++ test/specs/global.deprecations.tests.js
