+++ test/specs/platform.dom.tests.js
