+++ test/specs/scale.category.tests.js
