+++ test/specs/scale.logarithmic.tests.js
