+++ test/specs/scale.time.tests.js
